+++ src/AppBundle/Controller/Transaction/PurchaseReceiptHeaderController.php
@@ -19,7 +19,7 @@ class PurchaseReceiptHeaderController extends Controller
     /**
      * @Route("/grid", name="transaction_purchase_receipt_header_grid", condition="request.isXmlHttpRequest()")
      * @Method("POST")
-     * @Security("has_role('ROLE_TRANSACTION')")
+     * @Security("has_role('ROLE_PURCHASE_RECEIPT_HEADER_NEW') or has_role('ROLE_PURCHASE_RECEIPT_HEADER_EDIT') or has_role('ROLE_PURCHASE_RECEIPT_HEADER_DELETE')")
      */
     public function gridAction(Request $request)
     {
@@ -37,7 +37,7 @@ class PurchaseReceiptHeaderController extends Controller
     /**
      * @Route("/", name="transaction_purchase_receipt_header_index")
      * @Method("GET")
-     * @Security("has_role('ROLE_TRANSACTION')")
+     * @Security("has_role('ROLE_PURCHASE_RECEIPT_HEADER_NEW') or has_role('ROLE_PURCHASE_RECEIPT_HEADER_EDIT') or has_role('ROLE_PURCHASE_RECEIPT_HEADER_DELETE')")
      */
     public function indexAction()
     {
@@ -47,7 +47,7 @@ class PurchaseReceiptHeaderController extends Controller
     /**
      * @Route("/new.{_format}", name="transaction_purchase_receipt_header_new")
      * @Method({"GET", "POST"})
-     * @Security("has_role('ROLE_TRANSACTION')")
+     * @Security("has_role('ROLE_PURCHASE_RECEIPT_HEADER_NEW')")
      */
     public function newAction(Request $request, $_format = 'html')
     {
@@ -76,7 +76,7 @@ class PurchaseReceiptHeaderController extends Controller
     /**
      * @Route("/{id}", name="transaction_purchase_receipt_header_show", requirements={"id": "\d+"})
      * @Method("GET")
-     * @Security("has_role('ROLE_TRANSACTION')")
+     * @Security("has_role('ROLE_PURCHASE_RECEIPT_HEADER_NEW') or has_role('ROLE_PURCHASE_RECEIPT_HEADER_EDIT') or has_role('ROLE_PURCHASE_RECEIPT_HEADER_DELETE')")
      */
     public function showAction(PurchaseReceiptHeader $purchaseReceiptHeader)
     {
@@ -88,7 +88,7 @@ class PurchaseReceiptHeaderController extends Controller
     /**
      * @Route("/{id}/edit.{_format}", name="transaction_purchase_receipt_header_edit", requirements={"id": "\d+"})
      * @Method({"GET", "POST"})
-     * @Security("has_role('ROLE_TRANSACTION')")
+     * @Security("has_role('ROLE_PURCHASE_RECEIPT_HEADER_EDIT')")
      */
     public function editAction(Request $request, PurchaseReceiptHeader $purchaseReceiptHeader, $_format = 'html')
     {
@@ -117,15 +117,16 @@ class PurchaseReceiptHeaderController extends Controller
     /**
      * @Route("/{id}/delete", name="transaction_purchase_receipt_header_delete", requirements={"id": "\d+"})
      * @Method({"GET", "POST"})
-     * @Security("has_role('ROLE_TRANSACTION')")
+     * @Security("has_role('ROLE_PURCHASE_RECEIPT_HEADER_DELETE')")
      */
     public function deleteAction(Request $request, PurchaseReceiptHeader $purchaseReceiptHeader)
     {
+        $purchaseReceiptHeaderService = $this->get('app.transaction.purchase_receipt_header_form');
         $form = $this->createFormBuilder()->getForm();
         $form->handleRequest($request);
 
         if ($form->isSubmitted()) {
-            if ($form->isValid()) {
+            if ($form->isValid() && $purchaseReceiptHeaderService->isValidForDelete($purchaseReceiptHeader)) {
                 $purchaseReceiptHeaderService->delete($purchaseReceiptHeader);
 
                 $this->addFlash('success', array('title' => 'Success!', 'message' => 'The record was deleted successfully.'));
